==> 00_scraped_content_check.ipynb <==
#bring the name of the files in a folder
import os
import re

def get_files(folder):
    files = []
    for root, dirs, file in os.walk(folder):
        for f in file:
            files.append(f)
    return files

scraped_content = get_files('data/lovesac/scraped_content/')
#remove the extension of the files
scraped_content = [re.sub(r'\.md$', '', file) for file in scraped_content]
scraped_content


#read array from a text file
def read_array(file):
    with open(file, 'r') as f:
        return f.read().splitlines()

sitemap_urls = read_array('data/lovesac/sitemap_urls.txt')
#remain only the part after the last slash
#sitemap_urls = [re.sub(r'.*/', '', url) for url in sitemap_urls]
sitemap_urls

#compare two arrays, the first one is the one that should be in the second one
def compare_arrays(arr1, arr2):
    return [item for item in arr1 if item not in arr2]

compare_arrays(sitemap_urls, scraped_content)
