==> 0_extract_urls_from_sitemap.ipynb <==
import requests
import xml.etree.ElementTree as ET

sitemap_urls = ['https://www.lovesac.com/sitemap-content.xml' , 'https://www.lovesac.com/media/sitemap/sitemap.xml' ]
arr = []
for url in sitemap_urls:
    response = requests.get(url)
    sitemap_content = response.content

    root = ET.fromstring(sitemap_content)
    namespace = {'ns': 'http://www.sitemaps.org/schemas/sitemap/0.9'}

    urls = [url_elem.text for url_elem in root.findall('ns:url/ns:loc', namespaces=namespace)]
    arr.extend(urls)
arr
#array to text file
with open('data/lovesac/sitemap_urls.txt', 'w') as f:
    for item in arr:
        f.write("%s\n" % item)


#Broken URLs
import json
from pandas import json_normalize

file_path = 'data/lovesac/lovesac_brokenbacklinks_20240703.json'

with open(file_path, 'r') as file:
    data = json.load(file)

data = json_normalize(data['brokenBacklinks'])

#Save as txt file data['url_to'].values
with open('data/lovesac/broken_urls.txt', 'w') as f:
    for url in data['url_to'].values:
        f.write(url + '\n')
