==> 3_content_based_similarity_check.ipynb <==
import pandas as pd
from src.utils import Utils

utils = Utils()

#One time run to calculate and save embeddings
#document_dir = "data/lovesac/scraped_content"
#collection_name = 'lovesac'
#utils.calculate_save_embeddings(document_dir, collection_name)

#Read the broken backlinks and sitemap URLs from the files
with open('data/lovesac/broken_urls.txt') as f:
    broken_backlinks = f.read().splitlines()

with open('data/lovesac/sitemap_urls.txt') as f:
    sitemap_urls = f.read().splitlines()

broken_backlink = broken_backlinks[1]
df = utils.search_for_similar(broken_backlink, 'lovesac', k=3)  
df.insert(0, 'Broken Back Link', broken_backlink) 
df
